requirements.txt
discord.py
python-dotenv
pytest
requests-mock
